ecs_with_privatelink_vpc/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.27.0
aws-cdk.aws-apigateway==1.27.0
aws-cdk.aws-applicationautoscaling==1.27.0
aws-cdk.aws-autoscaling==1.27.0
aws-cdk.aws-autoscaling-common==1.27.0
aws-cdk.aws-autoscaling-hooktargets==1.27.0
aws-cdk.aws-certificatemanager==1.27.0
aws-cdk.aws-cloudformation==1.27.0
aws-cdk.aws-cloudfront==1.27.0
aws-cdk.aws-cloudwatch==1.27.0
aws-cdk.aws-ec2==1.27.0
aws-cdk.aws-ecr==1.27.0
aws-cdk.aws-ecr-assets==1.27.0
aws-cdk.aws-ecs==1.27.0
aws-cdk.aws-elasticloadbalancing==1.27.0
aws-cdk.aws-elasticloadbalancingv2==1.27.0
aws-cdk.aws-events==1.27.0
aws-cdk.aws-iam==1.27.0
aws-cdk.aws-kms==1.27.0
aws-cdk.aws-lambda==1.27.0
aws-cdk.aws-logs==1.27.0
aws-cdk.aws-route53==1.27.0
aws-cdk.aws-route53-targets==1.27.0
aws-cdk.aws-s3==1.27.0
aws-cdk.aws-s3-assets==1.27.0
aws-cdk.aws-sam==1.27.0
aws-cdk.aws-secretsmanager==1.27.0
aws-cdk.aws-servicediscovery==1.27.0
aws-cdk.aws-sns==1.27.0
aws-cdk.aws-sns-subscriptions==1.27.0
aws-cdk.aws-sqs==1.27.0
aws-cdk.aws-ssm==1.27.0
aws-cdk.core==1.27.0
aws-cdk.cx-api==1.27.0
aws-cdk.region-info==1.27.0
boto3==1.12.14
botocore==1.15.14
cattrs==1.0.0
docutils==0.15.2
importlib-metadata==1.5.0
importlib-resources==1.2.0
jmespath==0.9.5
jsii==0.22.0
pkg-resources==0.0.0
publication==0.0.3
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
zipp==3.1.0
